+++ lib/tableficate/filters/radio.rb
@@ -1,0 +1,6 @@
+module Tableficate
+  module Filter
+    class Radio < Base
+    end
+  end
+end