+++ spec/helper_spec.rb
@@ -1,4 +1,5 @@
 require 'spec_helper'
+include Tableficate::Helper
 
 describe Tableficate::Helper, type: :request do
   describe 'tableficate_select_tag' do
@@ -24,7 +25,7 @@ describe Tableficate::Helper, type: :request do
       visit '/filters/select_from_hash'
       page.should have_xpath("//select[@id='nobel_prize_winners_filter_category']/option[@value='Peace'][text()='3']")
     end
-    it 'selects an option from the params when the collection is a range' do
+    it 'selects an option from the params when the collection is a hash' do
       visit '/filters/select_from_hash?nobel_prize_winners[filter][category]=Peace'
       page.should have_xpath("//select[@id='nobel_prize_winners_filter_category']/option[@value='Peace'][@selected='selected']")
     end
@@ -48,43 +49,42 @@ describe Tableficate::Helper, type: :request do
     end
   end
 
-  describe 'tableficate_radio_tag' do
-    it 'allows for a collection to accept a range' do
-      visit '/filters/radio_from_range'
-      page.should have_xpath("//input[@type='radio'][@id='nobel_prize_winners_filter_year_1901'][@value='1901']")
-    end
-    it 'selects an option from the params when the collection is a range' do
-      visit '/filters/radio_from_range?nobel_prize_winners[filter][year]=1903'
-      page.should have_xpath("//input[@type='radio'][@id='nobel_prize_winners_filter_year_1903'][@value='1903'][@checked='checked']")
+  describe 'tableficate_radio_tags' do
+    it 'takes a block for custom output' do
+      visit '/filters/radio_tags?theme=custom_radio_block'
     end
 
-    it 'allows for a collection to accept a hash' do
-      visit '/filters/radio_from_hash'
-      page.should have_xpath("//label[@for='nobel_prize_winners_filter_category_Peace'][text()='3']")
-      page.should have_xpath("//input[@type='radio'][@id='nobel_prize_winners_filter_category_Peace'][@value='Peace']")
-    end
-    it 'selects an choice from the params when the collection is a range' do
-      visit '/filters/radio_from_hash?nobel_prize_winners[filter][category]=Peace'
-      page.should have_xpath("//input[@type='radio'][@id='nobel_prize_winners_filter_category_Peace'][@value='Peace'][@checked='checked']")
+    it 'should display a group of radio tags with no selection and nothing filtered' do
+      visit '/filters/radio_tags'
+
+      ['Chemistry', 'Literature', 'Peace', 'Physics', 'Physiology or Medicine'].each do |category|
+        id = "nobel_prize_winners_filter_category_#{category.gsub(/ /, '_')}"
+
+        page.should have_xpath("//input[@type='radio'][@id='#{id}'][@value='#{category}']")
+
+        page.has_no_checked_field?(id).should be true
+      end
+
+      page.should have_xpath('//table/tbody/tr', count: NobelPrizeWinner.joins(:nobel_prizes).size)
     end
 
-    it 'allows for a collection to accept an array' do
-      visit '/filters/radio_from_array'
-      page.should have_xpath("//input[@type='radio'][@id='nobel_prize_winners_filter_category_Peace'][@value='Peace']")
-    end
-    it 'selects an choice from the params when the collection is an array' do
-      visit '/filters/radio_from_array?nobel_prize_winners[filter][category]=Peace'
-      page.should have_xpath("//input[@type='radio'][@id='nobel_prize_winners_filter_category_Peace'][@value='Peace'][@checked='checked']")
-    end
+    it 'should display a group of radio tags with a selection and the table filtered' do
+      selected_category = 'Peace'
+      visit "/filters/radio_tags?nobel_prize_winners[filter][category]=#{selected_category}"
 
-    it 'allows for a collection to accept a nested array' do
-      visit '/filters/radio_from_nested_array'
-      page.should have_xpath("//label[@for='nobel_prize_winners_filter_category_Peace'][text()='3']")
-      page.should have_xpath("//input[@type='radio'][@id='nobel_prize_winners_filter_category_Peace'][@value='Peace']")
-    end
-    it 'selects an choice from the params when the collection is a nested array' do
-      visit '/filters/radio_from_nested_array?nobel_prize_winners[filter][category]=Peace'
-      page.should have_xpath("//input[@type='radio'][@id='nobel_prize_winners_filter_category_Peace'][@value='Peace'][@checked='checked']")
+      ['Chemistry', 'Literature', 'Peace', 'Physics', 'Physiology or Medicine'].each do |category|
+        id = "nobel_prize_winners_filter_category_#{category.gsub(/ /, '_')}"
+
+        page.should have_xpath("//input[@type='radio'][@id='#{id}'][@value='#{category}']")
+
+        if category == selected_category
+          page.has_checked_field?(id).should be true
+        else
+          page.has_no_checked_field?(id).should be true
+        end
+      end
+
+      page.should have_xpath('//table/tbody/tr', count: NobelPrizeWinner.joins(:nobel_prizes).where('nobel_prizes.category = ?', selected_category).size)
     end
   end
 end